ALE conserve KE
This commit adds two new diagnostics, and implements an update to
the ALE remap of velocity. The diagnostics compute the rate of
change of the depth integral of u^2/2 (ale_u2) and v^2/2 (ale_v2).
The update to the ALE remap of velocity corrects the remapped
velocity so that the kinetic energy is the same before and after
remapping. The correction preserves the order of the remapping
scheme: if the remapped velocity is second-order accurate before
the correction, then it is still second-order accurate after the
correction. The correction is activated by setting

REMAP_VEL_CONSERVE_KE = True

in the input namelist.